+++ NEWS.md
@@ -2,7 +2,8 @@
 
 * Fixed: `getEICs()` output was returned as nested matrices when `output="fill"`
 * Fixed: `report()`: component table lacked ret and m/z columns
-* Fixed: `newProject()`: misc fixes when creating a project that loads analyses from a table (issue #154). 
+* Fixed: `newProject()`: misc fixes when creating a project that loads analyses from a table (issue #154).
+* Fixed: Improve handling of raw data without exported MS2 isolation windows in `piek` and `generateMSPeakLists()` (issue #156)
 
 
 # patRoon 3.0